pipeline/requirements.txt
requests
pandas
boto3
pyarrow
python-dotenv